Configurable virtual component emission
Since the introduction of virtual components in adventure 4.18, mini
message color changing tags like gradient or rainbow have been making
use of them to store metadata in the parsed component with the goal of
reconstructing the color changing tag when deserialising.

As virtual components are however simply flattened into empty text
component when interacting with a serialiser that is not aware of them,
virtual components can quickly break the expected component layout of
the parsed component, specifically in relation to components parsed in
previous versions of mini message.

The commit introduces a simple flag on the MM builder that allows
consumers to disable the emission of virtual components.